Adds gitignore for media dir and EV spreadsheet
I'm going to use media directory for pictures moving forward. This
directory is in the gitignore to avoid taking up unnecessary storage on
github. I've also added an EV spreadsheet that needs a summary of what
it is used for typed in. I'm also planning to move forward with the
markdown conversion for the notes so that it can be used more easily
through github.